Feature: codebase refactoring
