+++ .gitignore
@@ -6,3 +6,4 @@ build/
 __pycache__/
 *.py[cod]
 version
+.env
