+++ .jujuignore
@@ -1,3 +1,4 @@
 /venv
 *.py[cod]
 *.charm
+.env
